requirements.txt
pandas==2.2.3
chardet==5.2.0
numpy==2.1.1
scikit-learn==1.5.2
matplotlib==3.9.2
json-numpy==2.1.0
requests==2.32.3
redis==5.1.0
boto3==1.35.29
py-mon==2.0.5
pymongo==4.9.1
python-dotenv==1.0.1
scipy==1.14.1
